misc, hook annotations wip
