+++ summoner/client/__init__.py
@@ -1,5 +1,5 @@
 """
-TODO: doc client and ClientMerger summary
+TODO: doc client, ClientMerger and ClientTranslation summary
 """
 from .client import SummonerClient
 from .merger import ClientMerger, ClientTranslation
