+++ summoner/server/__init__.py
@@ -1,1 +1,4 @@
+"""
+TODO: doc server
+"""
 from .server import SummonerServer
